+++ lib/swipable_stack.dart
@@ -1,4 +1,3 @@
 library swipable_stack;
 
 export 'src/swipable_stack.dart';
-export 'src/swipable_stack_controller.dart';
