+++ pubspec.yaml
@@ -1,10 +1,9 @@
 name: swipable_stack
 description: A new Flutter package project.
-version: 0.0.1
+version: 0.1.0-nullsafety.0
 homepage: https://github.com/HeavenOSK/flutter_swipable_stack
 environment:
   sdk: ">=2.12.0-0 <3.0.0"
-
 dependencies:
   flutter:
     sdk: flutter
